requirements.txt
fastapi
uvicorn
supabase
python-dotenv
PyPDF2
pdfplumber
pyjwt
requests
pydantic
reportlab
python-multipart